Prevent popwin from reusing window
When popwin tries to display a buffer, it will reuse an existing
window displaying the buffer if popwin:reuse-window is true. It is
annoying: for example when the *Help* buffer is displayed in a popup
window, clicking on a link will close the popup window and display the
*Help* buffer in a new window not handled by popwin. I don't want that
so I disable this behaviour.