Improve my/list-buffers-with-mode
Add a way to transform what is returned by the function. By default it
returns a list of buffers but if a function is passed as the second
argument it calls the function with the buffer as the sole argument
and what is returned is added to the result list.